+++ frontend/src/pages/options/scripts/dashboard/displayLoginStatus.js
