+++ frontend/src/pages/options/scripts/vocabulary/displayWordsLearned.js
