+++ frontend/src/pages/options/scripts/vocabulary/handleGameRedirect.js
@@ -1,0 +1,7 @@
+export default function handleGameRedirect() {
+    // Game redirect
+    const gameBtn = document.querySelector('.game-btn')
+    gameBtn.addEventListener('click', function () {
+        window.open('../game/index.html', '_blank');
+    })
+}
