+++ frontend/src/pages/options/scripts/vocabulary/index.js
@@ -1,0 +1,9 @@
+import handleGameRedirect from "./handleGameRedirect.js";
+import handleShowTable from "./handleShowTable.js";
+import popupTogle from "./table/popup/popupToggle.js";
+
+export default function vocabInit() {
+    handleShowTable();
+    handleGameRedirect();
+    popupTogle();
+}
